# Optimisation of compilation speed with CMake/Ninja