--- a/core/file/ofstream.cpp
+++ b/core/file/ofstream.cpp
@@ -16,6 +16,7 @@
 
 #include "file/ofstream.h"
 
+#include "exception.h"
 #include "file/utils.h"
 
 namespace MR::File {