--- a/src/dwi/tractography/roi.cpp
+++ b/src/dwi/tractography/roi.cpp
@@ -118,3 +118,5 @@ Image<bool> Mask::__get_mask(const std::string &name) {
 }
 
 } // namespace MR::DWI::Tractography
+
+template MR::Transform::Transform(const MR::DWI::Tractography::Mask &);