… .github/workflows/checks.yml
… CMakeLists.txt
… cmd/CMakeLists.txt
… core/CMakeLists.txt
… core/file/utils.h
… src/CMakeLists.txt
… src/dwi/tractography/roi.h
… src/gui/gui_pch.h
… src/registration/nonlinear.cpp
… src/registration/nonlinear.h
… src/registration/transform/base.cpp
… src/registration/transform/base.h
… src/registration/transform/reorient.cpp
… src/registration/transform/reorient.h