This PR aims to significantly reduce the time to compile MRtrix3 and address the issue raised in #2876.

I have tested the changes locally on my laptop on Fedora Asahi and Mac OS 14 and, roughly speaking, I see an improvement of almost 30% compared to `dev` (without using `ccache`).

Here is a table showing indicative build times (in seconds) on my Macbook M2 Pro 2023 (I took the best of 3 successive builds).

|                         | dev  | new  |
|-------------------------|------|------|
| MacOS 14 (AppleClang)   | 124s | 89s  |
| Fedora Asahi (clang 17) | 139s | 91s  |
| Fedora Asahi (gcc 13)   | 154s | 110s |

This was achieved with a lot of trial and error and the help of `ClangBuildAnalyzer`. @MRtrix3/mrtrix3-devs it'd be great if you could test this on your machines to see if you can reproduce similar results. One caveat is that to achieve the most optimal results, you must use the latest version of `Ninja`(v 1.12), released last week (see first point below). On MacOS, it's just a matter of running `brew upgrade`. On Linux you will probably need to download it directly from their [Github](https://github.com/ninja-build/ninja/releases) releases page and then run `export PATH=/ninja_dir/:$PATH` (it's unlikely that your distro has packaged the latest version). Before rebuilding, you obviously need to clear your build directory and compiler cache (`ccache -C`).

More in detail, there are three changes in this PR:

- Increased build parallelism. During the initial stages of our CMake effort, it was pointed out by @jdtournier that the CPU load during the build phase didn't stay 100% throughout its entirety. So looking more into this, I found that when building `dev`, the Ninja build graph looks like this:
<img width="1726" alt="Screenshot 2024-04-15 at 01 48 55" src="https://github.com/MRtrix3/mrtrix3/assets/129752264/ff6c3c47-9eef-47fd-8248-c8cb8d6b02dd"> If you look at the end of the graph, you can see that `mrregister.cpp.o` takes a long time to compile (about 45s) and in fact the rest of the object files finish compiling much earlier. So there is a period when the compiler is effectively operating using a single thread. Ideally, the compilation of `mrregister` should start at the beginning of the compilation phase, however, there is effectively no way to control the order of compilation in CMake (or in Ninja). To get around this, I used a simple trick (see `cmd/CMakeLists.txt`): create a dummy executable and then add `mrregister` as a dependency of this dummy ([this](https://github.com/ninja-build/ninja/pull/2177) is the relevant change in Ninja 1.12 that makes this possible). This is enough to trick Ninja and change the build graph to the following:
<img width="1726" alt="Screenshot 2024-04-15 at 01 53 52" src="https://github.com/MRtrix3/mrtrix3/assets/129752264/b0c9fa98-c140-4af4-ac42-77086d3fec37"> Now, you see that we are operating multi-threaded throughout the entire compilation phase. 

- Precompiled headers. Precompiled headers speed up compilation by storing parsed header information in a special file. During subsequent compilations, the compiler reuses this precompiled data instead of re-parsing headers. They are particularly suited for "stable" header files, i.e. files that do not change very often (e.g. STL headers). If files listed in a precompiled header file change, then a target using the precompiled header will need to be recompiled from scratch. In this PR, I only included the following headers (other than STL and third parties) in precompiled headers:  `gl.h` and `exception.h`. The first one is only used for `mrview/shview` and the second one is for all the code. If a dev needs to frequently edit these files, it'd be best to turn off `MRTRIX_USE_PCH` as otherwise, any change in those files will trigger a rebuild of the whole project.

- Explicit template instantiations. C++11 introduced the `extern template` syntax, which lets you say, "Hey, the template's definition? It's elsewhere!". So if we use this in a header file, we can tell the compiler to avoid automatically instantiating a specific template in *every* .cpp file that includes the header (of course we have to make sure that the template is instantiated somewhere, e.g. in the corresponding .cpp for the header).
After looking at the report by `ClangBuildAnalyzer`, I used this feature to explicitly instantiate some class and function templates to reduce compilation time. I did this fairly conservatively (I really don't like dealing with the complexity that templates bring), but the few changes I made still made a significant difference in build time.