@daljit46 2024-04-15
Managed to obtain further improvements by enabling PCHs for executables too (previously they were only used for the shared libraries.
The new timings look like below

|                         | dev  | new  |
|-------------------------|------|------|
| MacOS 14 (AppleClang)   | 124s | 84s  |
| Fedora Asahi (clang 17) | 139s | 94s  |
| Fedora Asahi (gcc 13)   | 154s | 103s |