@Lestropie 2024-04-15
Are you able to glean insight into what it is that makes the `mrregister` / `mrtransform` compilation so long? That would potentially be the place to be looking into precompiled headers / precompiled template instantiations.