@daljit46 2024-04-16
> Are you able to glean insight into what it is that makes the `mrregister` / `mrtransform` compilation so long? That would potentially be the place to be looking into precompiled headers / precompiled template instantiations.

Looking at the output of `ClangBuildAnalyzer`, for `mrregister` it shows that only 20% of the time is spent in the compiler's front-end (parsing and instantiation templates), while the rest of the time the compiler is busy optimising the code.
Hence I think the best course of action would be to split this file into multiple files (or at least move parts of its content somewhere else) so that the compiler can operate multithreaded.