@daljit46 2024-04-16
Further finding on `mrregister`: on my desk PC (a 20-core Intel(R) Xeon(R) CPU E5-2687W v3 @ 3.10GHz), the compilation of `mrregister.cpp.o` takes more than the time required to build all other the object files in the project!