@Lestropie 2024-05-09
Windows 11 machine through MSYS2, 16 cores, Ninja 1.12 (already available on MSYS2), average of 3 runs:

|       | `dev`  | new  |
|-------|--------|------|
| GCC   |   202s | 118s |
| Clang |   183s | 106s |

Can't currently test on my work Linux laptop as it's partway through a script doing 72 `bedpostx` runs and I don't particularly want to interrupt it.

I agree regarding abuse of inlining / too many implementations in header files / would benefit from use of `extern template`. Chances are most of the benefits have already been obtained here, and there may be somewhat diminishing returns hunting for more. So I'm not sure it's worth delaying the merge. I will myself try to keep in mind whenever any given area of code is under modification to consider rectifying such things on a case-by-case basis.